refactor: move transport internals to _transport/ folder
_http.py, _sse.py, _chat_proxy.py were scattered at the package root.
Moved to superme_sdk/_transport/ with a __init__.py that re-exports
key symbols. Old root files replaced with backward-compat shims.

All import sites updated to use _transport.* directly.
Docs updated: api/chat.md and api/low_level.md point to new paths.
