+++ superme_sdk/_transport/__init__.py
@@ -1,0 +1,6 @@
+"""Low-level HTTP, SSE, and protocol transport internals."""
+
+from ._http import HttpMixin, _decode_jwt
+from ._sse import aiter_sse_lines, iter_sse_lines
+
+__all__ = ["HttpMixin", "_decode_jwt", "iter_sse_lines", "aiter_sse_lines"]
