+++ superme_sdk/services/aio/_groups.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import json
 from typing import Any, AsyncGenerator, Optional
 
-from ..._sse import aiter_sse_lines
+from ..._transport._sse import aiter_sse_lines
 
 
 class AsyncGroupsMixin:
